Moved `prisonerLocation()` into `offenderSearchUtils` module
Alongside the other related functions.
