+++ server/utils/utils.test.ts
@@ -9,12 +9,11 @@ import {
   possessive,
   kebabCase,
   nameOfPerson,
-  prisonerLocation,
   reversedNameOfPerson,
   yearsSince,
 } from './utils'
 import { andrew, barry, chris, donald, ernie, fred } from '../data/testData/offenderSearch'
-import { isInPrison, isOutside, isBeingTransferred } from './offenderSearchUtils'
+import { isInPrison, isOutside, isBeingTransferred, prisonerLocation } from './offenderSearchUtils'
 import { fakeClock } from '../testutils/fakeJestClock'
 
 describe('convert to title case', () => {
